refactor : 🔄 #150 要素をコンポーネント化する
